+++ test/run_linecablemodels_extension_tests.jl
@@ -1,0 +1,1 @@
+include("LineCableModelsExt_test.jl")